Refactor codebase
The layout of the code has changed from having
two 'util' modules to only having one.
Thus core.py has been added to replace game_logic.py and util.py.

Discerning the difference between util/game_logic files seemed unclear.
Some of the code in game_logic also seems like a
better fit in some classes.

A configurations file has also been added
to allow configurations to be shared across modules.

This was needed as some of the configurations
were/are used across core and main.
