+++ classes/baseobject.py
@@ -7,11 +7,15 @@ from abc import ABC, abstractproperty
 
 
 class BaseObject(ABC):
-
     @abstractproperty
     def IMG(self):
 
         NotImplemented
+
+    @property
+    def NO_IMG(self):
+
+        return None
 
     def __init_subclass__(cls, *args, **kwargs):
         """Specifications required by future SpaceInvader subclasses."""
@@ -23,3 +27,8 @@ class BaseObject(ABC):
     def __init__(self, x_cord, y_cord):
         self.x_cord = x_cord
         self.y_cord = y_cord
+
+    def blit(self, screen):
+        if self.IMG == self.NO_IMG:
+            raise ValueError(f"No image is set for the {type(self).__name__} object!")
+        screen.blit(self.IMG, (self.x_cord, self.y_cord))
